src/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
asyncpg==0.29.0
pydantic==2.5.0
python-dotenv==1.0.0
pyjwt==2.8.0
cryptography==41.0.7
bcrypt==4.1.2
numpy==1.26.4
redis==5.0.1
prometheus-client==0.19.0
paho-mqtt==1.6.1
confluent-kafka==2.3.0